Refactor Bloomy Client to replace Measurable with Scorecard operations
